refactor(commands): separate id command text of id slash command
