+++ src/commands/id/mod.rs
@@ -1,0 +1,2 @@
+pub mod slash_command;
+pub mod text_command;
