+++ src/commands/id/slash_command/mod.rs
@@ -1,0 +1,1 @@
+pub mod id;
