+++ src/commands/id/text_command/mod.rs
@@ -1,0 +1,1 @@
+pub mod id;
